moved bible videos crud from content crud to media crud
